requirements.txt
opencv-python==4.8.0.74
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
-f https://download.pytorch.org/whl/cu121
colorlog
numpy
Pillow
ipython
imageio
tqdm
scipy
lark
matplotlib
scikit-image
wandb
rtpt
scikit-learn
grad-cam
einops
stable_baselines3
gym
numpy
timm
transformers
accelerate>=0.26.0
bitsandbytes
ace_tools